refactor: split assignment type checks
