+++ src/types/checker/stmt_check/assignments/static_properties.rs
@@ -1,0 +1,305 @@
+use crate::errors::CompileError;
+use crate::parser::ast::{Expr, StaticReceiver};
+use crate::span::Span;
+use crate::types::{PhpType, TypeEnv};
+
+use super::super::super::Checker;
+
+struct StaticPropertyAssignmentTarget {
+    class_name: String,
+    declaring_class: String,
+    property_has_declared_type: bool,
+    prop_ty: PhpType,
+}
+
+pub(super) fn check_static_property_assign(
+    checker: &mut Checker,
+    receiver: &StaticReceiver,
+    property: &str,
+    value: &Expr,
+    span: Span,
+    env: &mut TypeEnv,
+) -> Result<(), CompileError> {
+    let val_ty = checker.infer_type(value, env)?;
+    let target = resolve_static_property_assignment_target(checker, receiver, property, span)?;
+
+    if target.property_has_declared_type {
+        checker.require_compatible_arg_type(
+            &target.prop_ty,
+            &val_ty,
+            span,
+            &format!("Static property {}::${}", target.class_name, property),
+        )?;
+    }
+
+    if !target.property_has_declared_type {
+        refine_static_property_assignment_type(
+            checker,
+            property,
+            &target.declaring_class,
+            &val_ty,
+        );
+    }
+    Ok(())
+}
+
+pub(super) fn check_static_property_array_push(
+    checker: &mut Checker,
+    receiver: &StaticReceiver,
+    property: &str,
+    value: &Expr,
+    span: Span,
+    env: &mut TypeEnv,
+) -> Result<(), CompileError> {
+    let val_ty = checker.infer_type(value, env)?;
+    let target = resolve_static_property_assignment_target(checker, receiver, property, span)?;
+    let updated_prop_ty = match target.prop_ty {
+        PhpType::Array(elem_ty) => {
+            if target.property_has_declared_type {
+                checker.require_compatible_arg_type(
+                    elem_ty.as_ref(),
+                    &val_ty,
+                    span,
+                    &format!("Static property {}::${}[]", target.class_name, property),
+                )?;
+                PhpType::Array(elem_ty)
+            } else if *elem_ty == val_ty {
+                PhpType::Array(elem_ty)
+            } else {
+                let merged_ty = checker
+                    .merge_array_element_type(&elem_ty, &val_ty)
+                    .unwrap_or(val_ty.clone());
+                PhpType::Array(Box::new(merged_ty))
+            }
+        }
+        PhpType::Int | PhpType::Void if !target.property_has_declared_type => {
+            PhpType::Array(Box::new(val_ty.clone()))
+        }
+        PhpType::Buffer(_) => {
+            return Err(CompileError::new(
+                span,
+                "buffer<T> does not support push; allocate with buffer_new<T>(len)",
+            ))
+        }
+        other => {
+            return Err(CompileError::new(
+                span,
+                &format!("Array push requires an array static property, got {}", other),
+            ))
+        }
+    };
+
+    if !target.property_has_declared_type {
+        update_static_property_type(
+            checker,
+            property,
+            &target.declaring_class,
+            updated_prop_ty,
+        );
+    }
+    Ok(())
+}
+
+pub(super) fn check_static_property_array_assign(
+    checker: &mut Checker,
+    receiver: &StaticReceiver,
+    property: &str,
+    index: &Expr,
+    value: &Expr,
+    span: Span,
+    env: &mut TypeEnv,
+) -> Result<(), CompileError> {
+    let idx_ty = checker.infer_type(index, env)?;
+    let val_ty = checker.infer_type(value, env)?;
+    let target = resolve_static_property_assignment_target(checker, receiver, property, span)?;
+    if idx_ty != PhpType::Int {
+        return Err(CompileError::new(span, "Array index must be integer"));
+    }
+
+    let updated_prop_ty = match target.prop_ty {
+        PhpType::Array(elem_ty) => {
+            if target.property_has_declared_type {
+                checker.require_compatible_arg_type(
+                    elem_ty.as_ref(),
+                    &val_ty,
+                    span,
+                    &format!("Static property {}::${}[]", target.class_name, property),
+                )?;
+                PhpType::Array(elem_ty)
+            } else if *elem_ty == val_ty {
+                PhpType::Array(elem_ty)
+            } else {
+                let merged_ty = checker
+                    .merge_array_element_type(&elem_ty, &val_ty)
+                    .unwrap_or(val_ty.clone());
+                PhpType::Array(Box::new(merged_ty))
+            }
+        }
+        PhpType::Int | PhpType::Void if !target.property_has_declared_type => {
+            PhpType::Array(Box::new(val_ty.clone()))
+        }
+        other => {
+            return Err(CompileError::new(
+                span,
+                &format!(
+                    "Array index assignment requires an array static property, got {}",
+                    other
+                ),
+            ))
+        }
+    };
+
+    if !target.property_has_declared_type {
+        update_static_property_type(
+            checker,
+            property,
+            &target.declaring_class,
+            updated_prop_ty,
+        );
+    }
+    Ok(())
+}
+
+fn resolve_static_property_assignment_target(
+    checker: &Checker,
+    receiver: &StaticReceiver,
+    property: &str,
+    span: Span,
+) -> Result<StaticPropertyAssignmentTarget, CompileError> {
+    let class_name = match receiver {
+        StaticReceiver::Named(class_name) => class_name.as_str().to_string(),
+        StaticReceiver::Self_ => checker
+            .current_class
+            .as_ref()
+            .cloned()
+            .ok_or_else(|| CompileError::new(span, "Cannot use self:: outside class method scope"))?,
+        StaticReceiver::Static => checker
+            .current_class
+            .as_ref()
+            .cloned()
+            .ok_or_else(|| CompileError::new(span, "Cannot use static:: outside class method scope"))?,
+        StaticReceiver::Parent => {
+            let current_class = checker.current_class.as_ref().ok_or_else(|| {
+                CompileError::new(span, "Cannot use parent:: outside class method scope")
+            })?;
+            let current_info = checker.classes.get(current_class).ok_or_else(|| {
+                CompileError::new(span, &format!("Undefined class: {}", current_class))
+            })?;
+            current_info.parent.as_ref().cloned().ok_or_else(|| {
+                CompileError::new(
+                    span,
+                    &format!("Class {} has no parent class", current_class),
+                )
+            })?
+        }
+    };
+
+    let class_info = checker
+        .classes
+        .get(&class_name)
+        .ok_or_else(|| CompileError::new(span, &format!("Undefined class: {}", class_name)))?;
+    if !class_info
+        .static_properties
+        .iter()
+        .any(|(name, _)| name == property)
+    {
+        return Err(CompileError::new(
+            span,
+            &format!("Undefined static property: {}::{}", class_name, property),
+        ));
+    }
+    if let Some(visibility) = class_info.static_property_visibilities.get(property) {
+        let declaring_class = class_info
+            .static_property_declaring_classes
+            .get(property)
+            .map(String::as_str)
+            .unwrap_or(class_name.as_str());
+        if !checker.can_access_member(declaring_class, visibility) {
+            return Err(CompileError::new(
+                span,
+                &format!(
+                    "Cannot access {} static property: {}::{}",
+                    Checker::visibility_label(visibility),
+                    class_name,
+                    property
+                ),
+            ));
+        }
+    }
+    let declaring_class = class_info
+        .static_property_declaring_classes
+        .get(property)
+        .cloned()
+        .unwrap_or_else(|| class_name.clone());
+    let property_has_declared_type = class_info.declared_static_properties.contains(property);
+    let prop_ty = class_info
+        .static_properties
+        .iter()
+        .find(|(name, _)| name == property)
+        .map(|(_, ty)| ty.clone())
+        .unwrap_or(PhpType::Int);
+
+    Ok(StaticPropertyAssignmentTarget {
+        class_name,
+        declaring_class,
+        property_has_declared_type,
+        prop_ty,
+    })
+}
+
+fn update_static_property_type(
+    checker: &mut Checker,
+    property: &str,
+    declaring_class: &str,
+    updated_ty: PhpType,
+) {
+    for class_info in checker.classes.values_mut() {
+        if class_info
+            .static_property_declaring_classes
+            .get(property)
+            .map(String::as_str)
+            != Some(declaring_class)
+        {
+            continue;
+        }
+        if let Some(prop) = class_info
+            .static_properties
+            .iter_mut()
+            .find(|(name, _)| name == property)
+        {
+            prop.1 = updated_ty.clone();
+        }
+    }
+}
+
+fn refine_static_property_assignment_type(
+    checker: &mut Checker,
+    property: &str,
+    declaring_class: &str,
+    val_ty: &PhpType,
+) {
+    for class_info in checker.classes.values_mut() {
+        if class_info
+            .static_property_declaring_classes
+            .get(property)
+            .map(String::as_str)
+            != Some(declaring_class)
+        {
+            continue;
+        }
+        if let Some(prop) = class_info
+            .static_properties
+            .iter_mut()
+            .find(|(name, _)| name == property)
+        {
+            if matches!(prop.1, PhpType::Int | PhpType::Void) && prop.1 != *val_ty {
+                prop.1 = val_ty.clone();
+            } else {
+                let refined_ty = Checker::specialize_generic_array_hint(&prop.1, val_ty);
+                if refined_ty != prop.1 {
+                    prop.1 = refined_ty;
+                }
+            }
+        }
+    }
+}
